requirements.txt
fastapi
uvicorn
jinja2
telethon
pyrogram
pandas
openpyxl
python-dotenv
aiofiles
